Bug hunt. Command station would randomly crash, and produce erratic, perhaps damaging behavior in locomotives. Fixes included: Properly removing handlers from link upon deletion, increasing datagram timeout to 5 seconds (why so high? why isn't 2 enough? odd, and worrisome), and a few other odds and ends that unfortunately escape me just now. DEGW
git-svn-id: svn+ssh://svn.code.sf.net/p/openlcb/svn/trunk/prototypes/C@1320 87536cb5-86f1-4cdc-8fc6-3f0f71bcd267

+++ libraries/OpenLCB/OLCB_CAN_Link.h
@@ -93,7 +93,7 @@ class OLCB_CAN_Link : public OLCB_Link
   
   uint16_t getAlias();
  
-  unsigned long _aliasTimer, _aliasCacheTimer; // used to wait for specific times (Arduino type definition)
+  uint32_t _aliasTimer, _aliasCacheTimer; // used to wait for specific times (Arduino type definition)
   uint32_t lfsr1, lfsr2;  // PRNG sequence value: lfsr1 is upper 24 bits, lfsr2 lower
   uint8_t state; // internal state counter, starts at zero
